LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/claim-worker-fetch.https-expected.txt



Harness Error (TIMEOUT), message = null

PASS fetch() in Worker should be intercepted after the client is claimed.
FAIL fetch() in nested Worker should be intercepted after the client is claimed. assert_equals: fetch() should not be intercepted. expected "a simple text file\n" but got "Fail: undefined"
PASS fetch() in blob URL Worker should be intercepted after the client is claimed.
TIMEOUT fetch() in nested blob URL Worker created from a blob URL Worker should be intercepted after the client is claimed. Test timed out
NOTRUN fetch() in nested Worker created from a blob URL Worker should be intercepted after the client is claimed.
NOTRUN fetch() in nested blob URL Worker created from a Worker should be intercepted after the client is claimed.
